+++ src/Fluentify.Tests/Snippets/Classes.Global.cs
@@ -112,10 +112,14 @@ public static partial class Classes
                 Func<object, object> builder)
             {
                 subject.ThrowIfNull("subject");
-
                 builder.ThrowIfNull("builder");
 
-                var instance = new object();
+                var instance = subject.Attributes?.FirstOrDefault();
+
+                if (instance is null)
+                {
+                    instance = new object();
+                }
 
                 instance = builder(instance);