+++ src/Fluentify.Tests/Snippets/Classes.InvalidDescriptor.cs
@@ -122,10 +122,14 @@ public static partial class Classes
                     Func<object, object> builder)
                 {
                     subject.ThrowIfNull("subject");
-
                     builder.ThrowIfNull("builder");
 
-                    var instance = new object();
+                    var instance = subject.Attributes?.FirstOrDefault();
+
+                    if (instance is null)
+                    {
+                        instance = new object();
+                    }
 
                     instance = builder(instance);