+++ src/Fluentify.Tests/Snippets/Records.NestedInClassWithGenerics.cs
@@ -146,10 +146,14 @@ public static partial class Records
                     where TInner : struct
                 {
                     subject.ThrowIfNull("subject");
-
                     builder.ThrowIfNull("builder");
 
-                    var instance = new object();
+                    var instance = subject.Attributes?.FirstOrDefault();
+
+                    if (instance is null)
+                    {
+                        instance = new object();
+                    }
 
                     instance = builder(instance);