+++ src/Fluentify.Tests/Snippets/Records.SelfDescriptorOnRequired.cs
@@ -135,10 +135,14 @@ public static partial class Records
                     Func<object, object> builder)
                 {
                     subject.ThrowIfNull("subject");
-
                     builder.ThrowIfNull("builder");
 
-                    var instance = new object();
+                    var instance = subject.Attributes?.FirstOrDefault();
+
+                    if (instance is null)
+                    {
+                        instance = new object();
+                    }
 
                     instance = builder(instance);